API: Add lib.array_utils namespace
