+++ numpy/lib/tests/test_utils.py
@@ -1,40 +1,16 @@
 import pytest
 
 import numpy as np
-from numpy.testing import assert_equal, assert_raises_regex
+from numpy.testing import assert_raises_regex
 import numpy.lib._utils_impl as _utils_impl
 
 from io import StringIO
 
 
-class TestByteBounds:
-
-    def test_byte_bounds(self):
-        # pointer difference matches size * itemsize
-        # due to contiguity
-        a = np.arange(12).reshape(3, 4)
-        low, high = np.byte_bounds(a)
-        assert_equal(high - low, a.size * a.itemsize)
-
-    def test_unusual_order_positive_stride(self):
-        a = np.arange(12).reshape(3, 4)
-        b = a.T
-        low, high = np.byte_bounds(b)
-        assert_equal(high - low, b.size * b.itemsize)
-
-    def test_unusual_order_negative_stride(self):
-        a = np.arange(12).reshape(3, 4)
-        b = a.T[::-1]
-        low, high = np.byte_bounds(b)
-        assert_equal(high - low, b.size * b.itemsize)
-
-    def test_strided(self):
-        a = np.arange(12)
-        b = a[::2]
-        low, high = np.byte_bounds(b)
-        # the largest pointer address is lost (even numbers only in the
-        # stride), and compensate addresses for striding by 2
-        assert_equal(high - low, b.size * 2 * b.itemsize - b.itemsize)
+@pytest.mark.filterwarnings("ignore:.*safe_eval.*:DeprecationWarning")
+def test_safe_eval_nameconstant():
+    # Test if safe_eval supports Python 3.4 _ast.NameConstant
+    utils.safe_eval('None')
 
 
 def test_assert_raises_regex_context_manager():
