+++ numpy/typing/tests/data/pass/lib_utils.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 from io import StringIO
 
 import numpy as np
+import numpy.lib.array_utils as array_utils
 
 FILE = StringIO()
 AR = np.arange(10, dtype=np.float64)
@@ -12,8 +13,7 @@ def func(a: int) -> bool:
     return True
 
 
-np.byte_bounds(AR)
-np.byte_bounds(np.float64())
+array_utils.byte_bounds(AR)
+array_utils.byte_bounds(np.float64())
 
 np.info(1, output=FILE)
-
