+++ numpy/typing/tests/data/reveal/lib_utils.pyi
@@ -3,6 +3,7 @@ from io import StringIO
 
 import numpy as np
 import numpy.typing as npt
+import numpy.lib.array_utils as array_utils
 
 if sys.version_info >= (3, 11):
     from typing import assert_type
@@ -15,7 +16,7 @@ FILE: StringIO
 
 def func(a: int) -> bool: ...
 
-assert_type(np.byte_bounds(AR), tuple[int, int])
-assert_type(np.byte_bounds(np.float64()), tuple[int, int])
+assert_type(array_utils.byte_bounds(AR), tuple[int, int])
+assert_type(array_utils.byte_bounds(np.float64()), tuple[int, int])
 
 assert_type(np.info(1, output=FILE), None)
